bf16 cleanup and change fast_dtype to dtype

+++ torch/cuda/amp/autocast_mode.py
@@ -13,8 +13,8 @@ class autocast(torch.autocast_mode.autocast):
     See :class:`torch.autocast`.
     ``torch.cuda.amp.autocast(args...)`` is equivalent to ``torch.autocast("cuda", args...)``
     """
-    def __init__(self, enabled=True, fast_dtype=torch.float16):
-        super().__init__("cuda", enabled=enabled, fast_dtype=fast_dtype)
+    def __init__(self, enabled=True, dtype=torch.float16):
+        super().__init__("cuda", enabled=enabled, dtype=fast_dtype)
 
 
 # Casts Tensors and containers of Tensors.  Special-cases passthroughs for strings and np.ndarrays, which
